Bad apple day 10 compatibility
Did this instead of day 11 oops. Worth it. Was realllly hard to debug, I had to rebuild the algorithm like 3 times until I found that around line 8170 there are stalled cycles that actually use addx so it was hard for me to make an algorithm that processed that part properly.